Use stricter warning policy and update expected error message
libsdformat now emits an error when //inertial/pose/@relative_to is
non-empty, so this patch updates the error message to match the one
emitted by libsdformat. Support for //model/pose/@relative_to in Drake
is in the pipeline, but I kept the test that considers that invalid
until we merge #14401.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>