Add reusable configuration mechanism for malformed input
When a user asks us to parse untrustyworth data, we should allow them to
control how strict we should be.  If it was supposed to be error-free,
then we should allow them to fail-fast.  If they are just trying to bash
on regardless during protyping, then we should allow them to ignore.