[framework] Simplify scalar conversion implementation
Now that we've removed the use of traits dispatch during conversion,
we no longer need the sundry helper functions scatted around the header,
so we condense the implementation into a single Add and Convert function,
and move method definitions into the cc file.

This should be easier to understand and maintain, and allows us to remove
pydrake's duplication of the name-management logic; that is now internal
to the converter, not part of the functor capturing during Add.

This change marginally reduces the size of libdrake.so.