[pydrake] Use shared_ptr holder for Python classes with Clone
To simplify reference counting of Python classes that subclass a C++
base class that defines Clone (and thus need to override DoClone),
switch C++ base class bindings to use shared_ptr for its holder type
and add a DRAKE_OVERRIDE_PURE_CLONE macro with the logic for reference
counting when returning cloned objects from Python back to C++.

When using our current fork of pybind11 this is not strictly required,
but will be required when switching back to mainstream pybind11 that
lacks our bespoke unique_ptr holder tweaks.

SerializerInterface: For a given message type, it's OK to use one
const serializer across all systems that publish or subscribe for
efficiency. This also allows us to greatly simplify the Python
bindings.

Add missing bindings for deepcopy of any Trajectory (on the base
class) and remove the redundant bindings on each C++ subclass.