PR [symbolic] Give the default Variable a name
Remote various restriction on using the default Variable in an
Expression, Environment, Formula, etc.

This will be useful for univariate monomials and polynomials, where
the variable name doesn't matter but shouldn't be empty.

This is also probably a tiny speed-up to not hit the heap during the
default constructor (to allocate an empty reference-counted string).