PR [doc,tools] Adjust imports and libraries in prep for bzlmod
Re-spell imports to lose the "drake" part of the module namespace.
The old spelling of imports is dispreferred and will not work with
upcoming bzlmod.

For all Drake Python code, we should be using drake_py_library so that
all Drake-specific settings are always in effect. (Earlier in bzlmod
porting this was critically important; since then, I've changed some
options so that it's less critical but we should anyway follow this
hygiene so that we always have a single point of control for our build
rules.)