[trajectories] Add WrappedTrajectory and use in pydrake
The WrappedTrajectory is shallow-copy wrapper around any existing
Trajectory.

Because we allow implementations of Trajectory as Python subclasses,
we cannot assume that the deleter associated with a call to Clone is
`delete MyClass`, so we now adjust our PyTrajectory override logic to
wrap its return value in a WrappedTrajectory. The only reason the
unique_ptr used to work is Drake's custom fork or pybind11 with evil
hacks, which will be going away soon.

We also now allow Python implementations to implement the canonical
pythonic spelling of the "__deepcopy__" method, instead of the weird
Clone spelling.

We also move the binding for Clone onto the base class; it's not clear
why it was copy-pasted to all of the subclasses instead.