PR [wheel] Remove MUMPS, and other assorted repairs
Remove MUMPS support in wheel builds of Ipopt (breaking change).
This was already deprecated, but with some time still left on the
fuse. However, we were not installing MUMPS's subordinate licenses
for its suite of dynamic library dependencies, and since MUMPS is
deprecated anyway we might as well just remove it immediately
(in wheel builds), especially since SNOPT is always available.

For the remaining shared library dependencies in drake.libs, add code
to install their license notices as well as a new test that will fail
in case more dynamic libraries unexpectedly show up inside the wheel.

As part of this, notice that our dependency build of Reference-LAPACK
is dead code; we've been dynamically linking Ubuntu's default BLAS and
LAPACK into the wheel since forever, so remove it from dependencies.